src/main/webapp/app/entities/test-case/test-case-detail.component.ts
import { Component, OnInit } from '@angular/core';
import { ActivatedRoute } from '@angular/router';

import { ITestCase } from 'app/shared/model/test-case.model';

@Component({
  selector: 'jhi-test-case-detail',
  templateUrl: './test-case-detail.component.html'
})
export class TestCaseDetailComponent implements OnInit {
  testCase: ITestCase | null = null;

  constructor(protected activatedRoute: ActivatedRoute) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ testCase }) => (this.testCase = testCase));
  }

  previousState(): void {
    window.history.back();
  }
}
